// frontend/src/api/generated/model/thingModelProperty.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Nunu Example API
 * This is a sample server celler server.
 * OpenAPI spec version: 1.0.0
 */
import type { ThingModelPropertyValue } from './thingModelPropertyValue'

export interface ThingModelProperty {
  accessMode?: string
  dataType?: string
  identifier?: string
  name?: string
  reportedAt?: string
  unit?: string
  value?: ThingModelPropertyValue
}